+++ CHANGELOG.md
@@ -10,6 +10,7 @@
    insert.mjs.
  * Changed `onConflict` to receive the existing tree value as its first
    argument, rather than the node itself.
+ * Made `onConflict` for `insert` default to `THROW`. 
  * Removed `replaceWith`, which is no longer useful with the above change.
  * Removed the `notFoundAction` argument from `remove`.  You can detect
    whether the value to remove isn't found based on whether the tree is